[Refactor] Migrate help fuction to model/cmdtest
