+++ model/worker.go
@@ -1,9 +1,9 @@
 package model
 
 import (
-	"sync"
+	// "sync"
 	// "fmt"
-	"context"
+	// "context"
 	"github.com/sirupsen/logrus"
 )
 
@@ -11,25 +11,25 @@ var (
 	log = logrus.WithFields(logrus.Fields{"package": "model"})
 )
 
-type Worker interface {
-	// Start the worker with the given context
-	Start(context.Context) error
-	// Stop the worker
-	Stop() error
-	//   // Perform a job as soon as possibly
-	//   Perform(Job) error
-	//   // PerformAt performs a job at a particular time
-	//   PerformAt(Job, time.Time) error
-	//   // PerformIn performs a job after waiting for a specified amount of time
-	//   PerformIn(Job, time.Duration) error
-	//   // Register a Handler
-	//   Register(string, Handler) error
-}
-
-type BaseWorker struct {
-	startOnce  sync.Once
-	listenerId string
-	stop       chan bool
-	stopped    chan bool
-	jobs       chan Job
-}
+// type Worker interface {
+// 	// Start the worker with the given context
+// 	Start(context.Context) error
+// 	// Stop the worker
+// 	Stop() error
+// 	//   // Perform a job as soon as possibly
+// 	//   Perform(Job) error
+// 	//   // PerformAt performs a job at a particular time
+// 	//   PerformAt(Job, time.Time) error
+// 	//   // PerformIn performs a job after waiting for a specified amount of time
+// 	//   PerformIn(Job, time.Duration) error
+// 	//   // Register a Handler
+// 	//   Register(string, Handler) error
+// }
+//
+// type BaseWorker struct {
+// 	startOnce  sync.Once
+// 	listenerId string
+// 	stop       chan bool
+// 	stopped    chan bool
+// 	jobs       chan Job
+// }
