+++ worker/worker_test.go
@@ -1,0 +1,13 @@
+package worker
+
+// import (
+//
+// "github.com/weldpua2008/supraworker/job"
+// "github.com/weldpua2008/supraworker/model"
+//
+// )
+//
+// ctx, cancel := context.WithCancel(context.Background())
+// defer cancel() // cancel when we are getting the kill signal or exit
+// var wg sync.WaitGroup
+// jobs := make(chan model.Job, 1)
